refactor: 커스텀 훅 useVoteModal 리팩토링
- 파일 이름 변경: `useExecutionModal`
- 이미 존재하는 `useModal` 커스텀 훅 사용
